moving of channel message implementation
