+++ internal/blockchain/x/channel/msg/msg_server_deactivate_channel.go
@@ -1,4 +1,4 @@
-package keeper
+package msg
 
 import (
 	"context"
