+++ internal/blockchain/x/channel/msg/msg_server_how_is.go
@@ -1,4 +1,4 @@
-package keeper
+package msg
 
 import (
 	"context"
